pyproject.toml
[project]
name = "certwatchdog"
version = "0.1.0"
description = "SSL/TLS health checker with CLI and web UI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "httpx>=0.26.0",
    "pyyaml>=6.0",
    "pydantic>=2.5.0",
    "cryptography>=42.0.0",
    "certifi>=2023.0.0",
]

[project.scripts]
certwatchdog = "certwatchdog.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["certwatchdog"]

